Rename HandoutDocument -> Handout
We have to separate the current document:
 - Handout - the entity that is persisted
 - HandoutDocument - created by HandoutBehind and rendered

 https://github.com/Not-Endurance/not-timing-system/issues/254
