+++ src-v2/Judge/NTS.Judge.Blazor/Ports/IHandoutsBehind.cs
@@ -1,12 +1,12 @@
 ﻿using Not.Blazor.Ports.Behinds;
 using Not.Injection;
 using Not.Startup;
-using NTS.Domain.Core.Objects;
+using NTS.Domain.Core.Entities;
 
 namespace NTS.Judge.Blazor.Ports;
 
 public interface IHandoutsBehind : IStartupInitializer, IObservableBehind, ISingletonService
 {
-    IReadOnlyList<HandoutDocument> Handouts { get; }
-    Task<IEnumerable<HandoutDocument>> PopAll();
+    IReadOnlyList<Handout> Handouts { get; }
+    Task<IEnumerable<Handout>> PopAll();
 }
