+++ src-v2/NTS.Persistance/Adapters/HandoutRepository.cs
@@ -1,9 +1,9 @@
-﻿using NTS.Domain.Core.Objects;
+﻿using NTS.Domain.Core.Entities;
 using NTS.Persistence.States;
 
 namespace NTS.Persistence.Adapters;
 
-public class HandoutRepository : SetRepository<HandoutDocument, CoreState>
+public class HandoutRepository : SetRepository<Handout, CoreState>
 {
     public HandoutRepository(IStore<CoreState> store) : base(store)
     {
